Check if config exists

+++ thirdai_python_package/neural_db/neural_db.py
@@ -370,13 +370,11 @@ class NeuralDB:
 
         ray_version = ray.__version__
         if LooseVersion(ray_version) >= LooseVersion("2.7"):
-            warnings.warn(
-                """
+            warnings.warn("""
                 Using ray version 2.7 or higher requires specifying a remote or NFS storage path. 
                 Support for local checkpoints has been discontinued in these versions. 
                 Refer to https://github.com/ray-project/ray/issues/37177 for details.
-                """.strip()
-            )
+                """.strip())
 
         if not isinstance(documents, list) or not all(
             isinstance(doc, CSV) for doc in documents
@@ -571,8 +569,9 @@ class NeuralDB:
 
         on_success()
 
-        delete_folder(checkpoint_config.checkpoint_dir)
-        self.save(save_to=str(checkpoint_config.checkpoint_dir))
+        if checkpoint_config:
+            delete_folder(checkpoint_config.checkpoint_dir)
+            self.save(save_to=str(checkpoint_config.checkpoint_dir))
         return ids
 
     def delete(self, source_id: str):
